Rename DiffusionSampler→Diffusion (diffusion.py) and Diffusion→Agedi (agedi.py)
Agent-Logs-Url: https://github.com/nronne/agedi/sessions/7377a07e-a06d-4041-8c3f-5f1e7139e671
